Update messages to have command type as tag and iOS client time as payload. Add time synchronization and input buffer for more stable latency (very experimental).

+++ iOS/TaikoController/TaikoCommand.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-enum TaikoCommand : UInt8 {
+enum TaikoCommand : UInt32 {
     case LeftDon = 1
     case RightDon
     case LeftKa
